requirements.txt
img2pdf==0.4.3
lxml==4.7.1
packaging==21.3
pikepdf==4.4.1
Pillow==9.0.1
pyparsing==3.0.7
PyPDF2==1.26.0
